renamed configuration to config to match .cpp, dunno why all other components use configuration
