+++ Gems/AtomLyIntegration/CommonFeatures/Code/atomlyintegration_commonfeatures_public_files.cmake
@@ -13,7 +13,7 @@ set(FILES
     Include/AtomLyIntegration/CommonFeatures/CoreLights/DirectionalLightBus.h
     Include/AtomLyIntegration/CommonFeatures/CoreLights/DirectionalLightComponentConfig.h
     Include/AtomLyIntegration/CommonFeatures/Debug/RenderDebugBus.h
-    Include/AtomLyIntegration/CommonFeatures/Debug/RenderDebugComponentConfiguration.h
+    Include/AtomLyIntegration/CommonFeatures/Debug/RenderDebugComponentConfig.h
     Include/AtomLyIntegration/CommonFeatures/Decals/DecalBus.h
     Include/AtomLyIntegration/CommonFeatures/Decals/DecalComponentConfig.h
     Include/AtomLyIntegration/CommonFeatures/Decals/DecalConstants.h
